Adding new step7 functions
Added new functions to be done for step7 of the TENET analysis, involving downstream analyses. This upload covers functions for creation of scatterplots involving  expression of key genes and methylation of probes linked to them, circos plots of key genes to their linked enhancer probes, histograms of the top genes by the number of linked enhancer probes, and a survival script to generate survival analysis data and plots for the key genes and their linked probes.

Updated the DESCRIPTION to reflect additions of key packages. Also made a change to the documentation of the step6 top_tr_tabulation script to fix a typo on what the function outputs.