TENETR function refactor
Did a massive overhaul of functions. Most functions have not largely changed in function, with the exception of adding methodology to select different different methylation arrays for the data, and slightly changed the code for placing the legend in the survival plots. Other than that, existing functions have been updated with new names and documentation for them has been updated.