Update to step1 and step2 functions with argument defaults and bugfix to step5 directory creation
Added defaults to some arguments in the step1_make_external_datasets and step2_get_diffmeth_regions. Also changed the documentation slightly for the step2_get_diffmeth_regions to note that minexp should be set to a number less than the number of total tumor samples in the dataset. Also fixed a bug in step5 where a step5 directory was not created, preventing the creation of new files.