Testing TENETR refactored scripts
Uploaded test versions for steps 1,2,3,5, and several step7 functions to be tested. If these pass bug fixing they will replace the existing scripts for those functions. Main changes are to steps 1 and 2 to allow more options for the user to investigate promoter probes and also investigate probes that only overlap with a subset of the datasets. Also the output file for step2 includes lists of the indentified probes instead of entire methylation dataframes of their expression so downstream analyses had to change to incorporate this.