Bugfixes and additional documentation to step7_linked_probe_motif_searching
Fixed a bug with an incorrectly named variable which prevented function from working properly. I also removed testing code (which was commented out, so it didn't affect function, but does make the code cleaner), and updated the documentation so the examples provided can run without needing to load the necessary libraries.